New database
Creating new database table with column datatype specification and appending the data from worldometer to the database.
In future we will have to find the way to get data from worldometer and save it to the database by relating the rows from world_covid19 databse. and relate it to country_id table